Algoritmo de Bresenham e Classe MyPolygon

+++ AlgorithmsImplementation1/AlgorithmsImplementation1/Model/MyPoint.cs
@@ -8,16 +8,16 @@ namespace AlgorithmsImplementation1.Model
 {
     public class MyPoint
     {
-        public double m_X { get; set; }
-        public double m_Y { get; set; }
+        public int m_X { get; set; }
+        public int m_Y { get; set; }
 
         public MyPoint()
         {
-            this.m_X = -1.0;
-            this.m_Y = -1.0;
+            this.m_X = -1;
+            this.m_Y = -1;
         }
 
-        public MyPoint(double p_X, double p_Y)
+        public MyPoint(int p_X, int p_Y)
         {
             this.m_X = p_X;
             this.m_Y = p_Y;
